修复报类型转换警告的问题 增加从游戏菜单中返回游戏的按键

+++ MangoAPI/include/SpriteAnimation.h
@@ -33,7 +33,7 @@ private:
     sf::Texture m_texture;
     std::unordered_map<std::string, std::vector<sf::IntRect>> m_frames;
 
-    int m_currentIndex = 0;
+    size_t m_currentIndex = 0;
     float m_animSpeed = 0.f;
 
     std::string m_animName;
